Cemiterio/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.9
django-bootstrap-datepicker-plus==3.0.5
django-heroku==0.0.0
gunicorn==20.1.0
Pillow==8.4.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
whitenoise==5.3.0
psycopg2
django-crispy-forms==1.13.0
